Make separate_replies key a separate input param
Since separate_replies is actualy a config setting rather than part
of the search query params inside the Search class, pass it in as a
separate argument to avoid adding logic into the Search class. The
logic required to add it as part of the params would modify the params
multidict which is a little dangerous since if done incorrectly could
lead to a different search query.