Move first stage build asset include above copy ..
Previously, the build assets from the first stage weren't being
copied into the second stage until after the copy .. which means
that this build step would have been repeated every time as generaly
we rebuild when some of the files change. In the case where assets
have to be rebuilt it is advantageous to put this step as far down
in the second stage as possible to avoid uneccesary rebuilds in the
second stage. However, in the case where the code in the second stage
changes, triggering a rebuild of the second stage but not the first
stage assets, it's advantageous to place this as far up in the file
as possible. Placing the assets include right above the copy command
satisfies both of these criteria. The copy . . will always force a
re-build and steps above it won't be re-built unless there are external
library dependency changes which happen way less than front end and
python code changes.